// src/edgedb/generated/edgeql-js/insert.ts
export type pointerIsOptional<T extends PropertyDesc | LinkDesc> =
  T["cardinality"] extends
    | Cardinality.Many
    | Cardinality.Empty
    | Cardinality.AtMostOne
    ? true
    : false;

export type InsertShape<El extends ObjectType> = typeutil.flatten<
  RawInsertShape<El>
>;

export type RawInsertShape<El extends ObjectType> =
  // short-circuit infinitely deep
  ObjectType extends El
    ? never
    : typeutil.stripNever<
        stripNonInsertables<stripBacklinks<El["__pointers__"]>>
      > extends infer Shape
    ? Shape extends ObjectTypePointers
      ? typeutil.addQuestionMarks<{
          [k in keyof Shape]:
            | pointerToAssignmentExpression<Shape[k]>
            | (pointerIsOptional<Shape[k]> extends true
                ? undefined | null
                : never)
            | (Shape[k]["hasDefault"] extends true ? undefined : never);
        }> & {[k in `@${string}`]: TypeSet | scalarLiterals}
      : never
    : never;

interface UnlessConflict {
  on: TypeSet | null;
  else?: TypeSet;
}

type InsertBaseExpression<Root extends TypeSet = TypeSet> = {
  __kind__: ExpressionKind.Insert;
  __element__: Root["__element__"];
  __cardinality__: Cardinality.One;
  __expr__: stripSet<Root>;
  __shape__: any;
};
export type $expr_Insert<
  // Root extends $expr_PathNode = $expr_PathNode
  El extends ObjectType = ObjectType
  // Conflict = UnlessConflict | null
  // Shape extends InsertShape<Root> = any
> = Expression<{
  __kind__: ExpressionKind.Insert;
  __element__: El;
  __cardinality__: Cardinality.One;
  __expr__: $expr_PathNode;
  __shape__: InsertShape<El>;

  unlessConflict(): $expr_InsertUnlessConflict<
    El,
    // Expression<{
    //   __kind__: ExpressionKind.Insert;
    //   __element__: El;
    //   __cardinality__: Cardinality.One;
    //   __expr__: $expr_PathNode;
    //   __shape__: InsertShape<El>;
    // }>,
    {on: null}
  >;
  unlessConflict<Conflict extends UnlessConflict>(
    conflictGetter: (scope: $scopify<El>) => Conflict
  ): $expr_InsertUnlessConflict<
    El,
    // Expression<{
    //   __kind__: ExpressionKind.Insert;
    //   __element__: El;
    //   __cardinality__: Cardinality.One;
    //   __expr__: $expr_PathNode;
    //   __shape__: InsertShape<El>;
    // }>,
    Conflict
  >;
}>;

export type $expr_InsertUnlessConflict<
  El extends ObjectType = ObjectType,
  // Root extends InsertBaseExpression = InsertBaseExpression,
  Conflict extends UnlessConflict = UnlessConflict
> = Expression<{
  __kind__: ExpressionKind.InsertUnlessConflict;
  __element__: Conflict["else"] extends TypeSet
    ? Conflict["else"]["__element__"]["__name__"] extends El["__name__"]
      ? El
      : $Object
    : El;
  __cardinality__: Conflict["else"] extends TypeSet
    ? Conflict["else"]["__cardinality__"]
    : Cardinality.AtMostOne;
  __expr__: InsertBaseExpression;
  __conflict__: Conflict;
}>;