// src/edgedb/generated/edgeql-js/select.ts
export type OrderByExpr = TypeSet<ScalarType | ObjectType, Cardinality>;
export type OrderByObjExpr = {
  expression: OrderByExpr;
  direction?: OrderByDirection;
  empty?: OrderByEmpty;
};

export type OrderByExpression =
  | OrderByExpr
  | OrderByObjExpr
  | [OrderByExpr | OrderByObjExpr, ...(OrderByExpr | OrderByObjExpr)[]];

export type OffsetExpression = TypeSet<
  $number,
  Cardinality.Empty | Cardinality.One | Cardinality.AtMostOne
>;

export type SelectFilterExpression = TypeSet<$bool, Cardinality>;
export type LimitOffsetExpression = TypeSet<
  $number,
  Cardinality.Empty | Cardinality.One | Cardinality.AtMostOne
>;
export type LimitExpression = TypeSet<
  $number,
  Cardinality.Empty | Cardinality.One | Cardinality.AtMostOne
>;

export type SelectModifierNames =
  | "filter"
  | "filter_single"
  | "order_by"
  | "offset"
  | "limit";

export type exclusivesToFilterSingle<E extends ExclusiveTuple> =
  ExclusiveTuple extends E
    ? never
    : E extends []
    ? never
    : {
        [j in keyof E]: {
          [k in keyof E[j]]: E[j][k] extends ObjectTypeSet
            ? TypeSet<
                anonymizeObject<E[j][k]["__element__"]>,
                E[j][k]["__cardinality__"]
              >
            : orLiteralValue<E[j][k]>;
        };
      }[number];
export type SelectModifiers<T extends ObjectType = ObjectType> = {
  // export type SelectModifiers = {
  filter?: SelectFilterExpression;
  filter_single?: // | Partial<
  //     typeutil.stripNever<{
  //       [k in keyof T["__pointers__"]]: T["__pointers__"][k]
  // extends PropertyDesc
  //         ? orScalarLiteral<{
  //             __element__: T["__pointers__"][k]["target"];
  //             __cardinality__: T["__pointers__"][k]["cardinality"];
  //           }>
  //         : never;
  //     }>
  //   >

  // | (ObjectType extends T
  //       ? unknown
  //       : typeutil.stripNever<{
  //           [k in keyof T["__pointers__"]]: T["__pointers__"][k]
  // extends PropertyDesc<
  //             infer T,
  //             infer C,
  //             infer E
  //           >
  //             ? E extends true
  //               ? orScalarLiteral<{
  //                   __element__: T;
  //                   __cardinality__: C;
  //                 }>
  //               : never
  //             : never;
  //         }>)
  exclusivesToFilterSingle<T["__exclusives__"]> | SelectFilterExpression;

  // | (ObjectType extends T
  //     ? unknown
  //     : typeutil.stripNever<{
  //         [k in keyof T["__pointers__"]]: T["__pointers__"][k]
  // extends PropertyDesc<
  //           infer T,
  //           infer C,
  //           infer E
  //         >
  //           ? E extends true
  //             ? orScalarLiteral<{
  //                 __element__: T;
  //                 __cardinality__: C;
  //               }>
  //             : never
  //           : never;
  //       }>);
  order_by?: OrderByExpression;
  offset?: OffsetExpression | number;
  limit?: LimitExpression | number;
};

export type UnknownSelectModifiers = {[k in keyof SelectModifiers]: unknown};
// ...
export type InferOffsetLimitCardinality<
  Card extends Cardinality,
  Modifers extends UnknownSelectModifiers
> = Modifers["limit"] extends number | LimitExpression
  ? cardutil.overrideLowerBound<Card, "Zero">
  : Modifers["offset"] extends number | OffsetExpression
  ? cardutil.overrideLowerBound<Card, "Zero">
  : Card;

// export type ComputeSelectCardinality<
//   Expr extends ObjectTypeExpression,
//   Modifiers extends UnknownSelectModifiers
// > = InferOffsetLimitCardinality<
//   InferFilterCardinality<Expr, Modifiers["filter"]>,
//   Modifiers
// >;
export type ComputeSelectCardinality<
  Expr extends ObjectTypeExpression,
  Modifiers extends UnknownSelectModifiers
> = InferOffsetLimitCardinality<
  undefined extends Modifiers["filter_single"]
    ? Expr["__cardinality__"]
    : cardutil.overrideUpperBound<Expr["__cardinality__"], "One">,
  Modifiers
>;
// ...
export type linkDescToLinkProps<Desc extends LinkDesc> = {
  [k in keyof Desc["properties"] & string]: $expr_PathLeaf<
    TypeSet<
      Desc["properties"][k]["target"],
      Desc["properties"][k]["cardinality"]
    >
    // {
    //   type: $scopify<Desc["target"]>;
    //   linkName: k;
    // },
    // Desc["properties"][k]["exclusive"]
  >;
};

export type pointersToObjectType<P extends ObjectTypePointers> = ObjectType<
  string,
  P,
  {}
>;

type linkDescToShape<L extends LinkDesc> = objectTypeToSelectShape<
  L["target"]
> &
  objectTypeToSelectShape<pointersToObjectType<L["properties"]>> &
  SelectModifiers;
export type linkDescToSelectElement<L extends LinkDesc> =
  | boolean
  // | pointerToCastableExpression<Shape[k]>
  | TypeSet<
      anonymizeObject<L["target"]>,
      cardutil.assignable<L["cardinality"]>
    >
  | linkDescToShape<L>
  | ((
      scope: $scopify<L["target"]> & linkDescToLinkProps<L>
    ) => linkDescToShape<L>);

// object types -> pointers
// pointers -> links
// links -> target object type
// links -> link properties
export type objectTypeToSelectShape<T extends ObjectType = ObjectType> =
  // ObjectType extends T
  //   ? {[k: string]: unknown}
  //   :
  Partial<{
    [k in keyof T["__pointers__"]]: T["__pointers__"][k] extends PropertyDesc
      ?
          | boolean
          | TypeSet<
              T["__pointers__"][k]["target"],
              cardutil.assignable<T["__pointers__"][k]["cardinality"]>
            >
          | $expr_PolyShapeElement
      : T["__pointers__"][k] extends LinkDesc
      ? linkDescToSelectElement<T["__pointers__"][k]>
      : any;
  }> & {[k: string]: unknown};